#5:  Admin auto-install
Removes the separate SQL script.  The installation looks for the activation of the bootstrap template and adds its new configuration settings if not already present.  If any of the configuration values are different from the expected default, a log file (with SQL that can be copied and pasted) is generated and a message issued to the current admin.