#24: Initialization changes
Check for to-be-added/updated configuration settings when a SuperUser is logged in.  Present the admin with a message that identifies any newly-added configuration values.
When the store's template is changed via Tools->Template Selection, check to see if that template contains a file unique to "bootstrap" (or a clone) and verify the built-in Zen Cart settings, producing a log-file for the admin.